+++ src/apps/nem/mosaic/serialize.py
@@ -1,30 +1,6 @@
-from .writers import *
-from trezor.messages.NEMMosaic import NEMMosaic
-from trezor.messages import NEMSupplyChangeType
-from apps.nem.layout import *
-
-
-async def ask_mosaic_creation(ctx, msg: NEMSignTx):
-    await require_confirm_action(ctx, 'Create mosaic "' + msg.mosaic_creation.definition.mosaic + '" under  namespace "'
-                                 + msg.mosaic_creation.definition.namespace + '"?')
-    await require_confirm_properties(ctx, msg.mosaic_creation.definition)
-    await require_confirm_fee(ctx, 'Confirm creation fee', msg.mosaic_creation.fee)
-
-    await require_confirm_final(ctx, msg.transaction.fee)
-
-
-async def ask_mosaic_supply_change(ctx, msg: NEMSignTx):
-    await require_confirm_action(ctx, 'Modify supply for "' + msg.supply_change.mosaic + '" under  namespace "'
-                                 + msg.supply_change.namespace + '"?')
-    if msg.supply_change.type == NEMSupplyChangeType.SupplyChange_Decrease:
-        ask_msg = 'Decrease supply by ' + str(msg.supply_change.delta) + ' whole units?'
-    elif msg.supply_change.type == NEMSupplyChangeType.SupplyChange_Increase:
-        ask_msg = 'Increase supply by ' + str(msg.supply_change.delta) + ' whole units?'
-    else:
-        raise ValueError('Invalid supply change type')
-    await require_confirm_action(ctx, ask_msg)
-
-    await require_confirm_final(ctx, msg.transaction.fee)
+from apps.nem.writers import *
+from apps.nem.helpers import *
+from trezor.messages.NEMSignTx import NEMSignTx
 
 
 def serialize_mosaic_creation(msg: NEMSignTx, public_key: bytes):
@@ -96,45 +72,3 @@ def _write_property(w: bytearray, name: str, value):
     write_uint32(w, 4 + len(name) + 4 + len(value))
     write_bytes_with_length(w, bytearray(name))
     write_bytes_with_length(w, bytearray(value))
-
-
-def serialize_mosaic(w: bytearray, namespace: str, mosaic: str, quantity: int):
-    identifier_length = 4 + len(namespace) + 4 + len(mosaic)
-    # indentifier length (u32) + quantity (u64) + identifier size
-    write_uint32(w, 4 + 8 + identifier_length)
-    write_uint32(w, identifier_length)
-    write_bytes_with_length(w, bytearray(namespace))
-    write_bytes_with_length(w, bytearray(mosaic))
-    write_uint64(w, quantity)
-
-
-def canonicalize_mosaics(mosaics: list):
-    if len(mosaics) <= 1:
-        return mosaics
-    mosaics = merge_mosaics(mosaics)
-    return sort_mosaics(mosaics)
-
-
-def are_mosaics_equal(a: NEMMosaic, b: NEMMosaic) -> bool:
-    if a.namespace == b.namespace and a.mosaic == b.mosaic:
-        return True
-    return False
-
-
-def merge_mosaics(mosaics: list) -> list:
-    if not len(mosaics):
-        return list()
-    ret = list()
-    for i in mosaics:
-        found = False
-        for k, y in enumerate(ret):
-            if are_mosaics_equal(i, y):
-                ret[k].quantity += i.quantity
-                found = True
-        if not found:
-            ret.append(i)
-    return ret
-
-
-def sort_mosaics(mosaics: list) -> list:
-    return sorted(mosaics, key=lambda m: (m.namespace, m.mosaic))
