+++ src/apps/nem/multisig/serialize.py
@@ -1,39 +1,8 @@
-from .writers import *
-from apps.nem.layout import *
+from apps.nem.writers import *
+from apps.nem.helpers import *
+from trezor.messages.NEMSignTx import NEMSignTx
 from trezor.crypto import hashlib
-from trezor.messages import NEMModificationType
 from trezor.crypto import nem
-
-
-async def ask_multisig(ctx, msg: NEMSignTx):
-    address = nem.compute_address(msg.multisig.signer, msg.transaction.network)
-    if msg.cosigning:
-        await require_confirm_address(ctx, 'Cosign transaction for', address)
-    else:
-        await require_confirm_address(ctx, 'Initiate transaction for', address)
-    await require_confirm_fee(ctx, 'Confirm multisig fee', msg.multisig.fee)
-
-
-async def ask_aggregate_modification(ctx, msg: NEMSignTx):
-    if not msg.multisig:
-        await require_confirm_action(ctx, 'Convert account to multisig account?')
-
-    for m in msg.aggregate_modification.modifications:
-        if m.type == NEMModificationType.CosignatoryModification_Add:
-            action = 'Add'
-        else:
-            action = 'Remove'
-        address = nem.compute_address(m.public_key, msg.transaction.network)
-        await require_confirm_address(ctx, action + ' cosignatory?', address)
-
-    if msg.aggregate_modification.relative_change:
-        if not msg.multisig:
-            action = 'Set minimum cosignatories to '
-        else:
-            action = 'Modify the number of cosignatories by '
-        await require_confirm_action(ctx, action + str(msg.aggregate_modification.relative_change) + '?')
-
-    await require_confirm_final(ctx, msg.transaction.fee)
 
 
 def serialize_multisig(msg: NEMTransactionCommon, public_key: bytes, inner: bytes):
